source/lib/state_representation/python/requirements.txt
pytest
numpy
numpy-quaternion
numba
scipy
